[Refactor] Move helper functions to utils module
